redirect debug info to netconsole
Redirect if debug and netconsole are both set in command line options.
The redirecting code is added in both debug and netconsole entry to
make the redirection as early as possible.
Debug info saving to file feature is still available if 'debug' is set
but 'netconsole' is not set in command line options.

Original idea was from Ian's post:
http://www.hellion.org.uk/blog/posts/debugging-initramfs-over-netconsole/

Signed-off-by: Roger Shimizu <rogershimizu@gmail.com>
Signed-off-by: Ian Campbell <ijc@debian.org>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>