hook-functions: Include soft-dependencies of all kernel modules
The proper solution to hidden module dependencies is to add 'softdep'
information to the depending modules.  Unfortunately we currently
still won't see those dependencies as 'modprobe --ignore-install'
inhibits processing of both 'install' and 'softdep' configuration
lines and embedded softdep information.

There are good reasons for using --ignore-install (see #384043), so
instead of removing this option we run modprobe twice, with and
without it.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>