parse_numeric: Add support for 20-bit minor numbers in 'root' parameter
Linux supports 32-bit device numbers in the format 0xmmmMMMmm, where
each m is a hex digit of the minor number and each M is a hex digit of
the major number.

Currently parse_lilo only works if the top 3 digits are 0 (or absent).
Shift and mask the upper bits correctly.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>